pyproject.toml
[tool.poetry]
name = "django-todo-rest-application"
version = "0.1.0"
description = "Django Todo REST Application"
authors = ["Amogh Madan <amoghmadaan@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
Django = ">=3.2.0,<3.3.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
tox = "^3.26.0"


[tool.poetry.group.production.dependencies]
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
